Army size is Natural for that sweet sweet arithmetic underflow checking

+++ hs/app/Page/Replay.hs
@@ -87,7 +87,7 @@ toMap replay@Replay{..} commandEvent = do
       | index <- mountains
       ]
     citiesMap = fromList $
-      [ (toCoord index, City (Neutral `Army` size))
+      [ (toCoord index, City (Neutral `Army` fromIntegral size))
       | (index, size) <- zip cities cityArmies
       ]
     generalsMap = fromList $
